backend/requirements.txt
# =============================================================================
# Core Framework
# =============================================================================
Django==5.2.3
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0

# =============================================================================
# Database
# =============================================================================
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# =============================================================================
# File Storage (AWS S3)
# =============================================================================
boto3==1.34.0
django-storages==1.14.2

# =============================================================================
# API Documentation
# =============================================================================
drf-spectacular==0.28.0

# =============================================================================
# Google APIs (Calendar + Meet)
# =============================================================================
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.110.0

# =============================================================================
# Email Service (Brevo)
# =============================================================================
requests==2.31.0

# =============================================================================
# Task Queue (Celery + Redis) - Optional for later
# =============================================================================
# celery==5.3.4
# redis==5.0.1
# django-celery-beat==2.8.0

# =============================================================================
# Filters
# =============================================================================
django-filter==24.3

# =============================================================================
# CORS
# =============================================================================
django-cors-headers==4.3.1

# =============================================================================
# Environment Variables
# =============================================================================
python-decouple==3.8

# =============================================================================
# Production Server
# =============================================================================
gunicorn==21.2.0
whitenoise==6.6.0

# =============================================================================
# Testing
# =============================================================================
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
freezegun==1.4.0

# =============================================================================
# Linting & Formatting
# =============================================================================
ruff==0.1.9
